requirements.txt
fastapi[all]
sqlalchemy
psycopg2
passlib[bcrypt]
python-jose[cryptography]
aiogram
alembic